feat: segment budget timer (#374)
* chore: refactor RundownTiming components

* feat: allow for negative segment duration (WIP)
