+++ meteor/client/ui/RundownView/RundownTiming/RundownTimingProvider.tsx
@@ -2,116 +2,21 @@ import { Meteor } from 'meteor/meteor'
 import * as React from 'react'
 import * as PropTypes from 'prop-types'
 import * as _ from 'underscore'
-import { withTracker } from '../../lib/ReactMeteorData/react-meteor-data'
-import { Rundown } from '../../../lib/collections/Rundowns'
-import { Part, Parts, PartId } from '../../../lib/collections/Parts'
-import { getCurrentTime, literal, normalizeArray, unprotectString } from '../../../lib/lib'
-import { RundownUtils } from '../../lib/rundown'
-import { MeteorReactComponent } from '../../lib/MeteorReactComponent'
-import { RundownPlaylist } from '../../../lib/collections/RundownPlaylists'
-import ClassNames from 'classnames'
-import { SpeechSynthesiser } from '../../lib/speechSynthesis'
+import { withTracker } from '../../../lib/ReactMeteorData/react-meteor-data'
+import { Part, PartId } from '../../../../lib/collections/Parts'
+import { getCurrentTime, literal, unprotectString } from '../../../../lib/lib'
+import { MeteorReactComponent } from '../../../lib/MeteorReactComponent'
+import { RundownPlaylist } from '../../../../lib/collections/RundownPlaylists'
 import {
 	PartInstance,
 	findPartInstanceOrWrapToTemporary,
-	PartInstanceId,
 	wrapPartToTemporaryInstance,
-} from '../../../lib/collections/PartInstances'
-import { Settings } from '../../../lib/Settings'
+} from '../../../../lib/collections/PartInstances'
+import { Settings } from '../../../../lib/Settings'
+import { RundownTiming, TimeEventArgs } from './RundownTiming'
 
 // Minimum duration that a part can be assigned. Used by gap parts to allow them to "compress" to indicate time running out.
 const MINIMAL_NONZERO_DURATION = 1
-
-export interface TimeEventArgs {
-	currentTime: number
-}
-
-export type TimingEvent = CustomEvent<TimeEventArgs>
-
-export namespace RundownTiming {
-	/**
-	 * Events used by the RundownTimingProvider
-	 * @export
-	 * @enum {number}
-	 */
-	export enum Events {
-		/** Event is emitted every now-and-then, generally to be used for simple displays */
-		'timeupdate' = 'sofie:rundownTimeUpdate',
-		/** event is emitted with a very high frequency (60 Hz), to be used sparingly as
-		 * hooking up Components to it will cause a lot of renders
-		 */
-		'timeupdateHR' = 'sofie:rundownTimeUpdateHR',
-	}
-
-	/**
-	 * Context object that will be passed to listening components. The dictionaries use the Part ID as a key.
-	 * @export
-	 * @interface RundownTimingContext
-	 */
-	export interface RundownTimingContext {
-		/** This is the total duration of the rundown as planned (using expectedDurations). */
-		totalRundownDuration?: number
-		/** This is the content remaining to be played in the rundown (based on the expectedDurations).  */
-		remainingRundownDuration?: number
-		/** This is the total duration of the rundown: as planned for the unplayed (skipped & future) content, and as-run for the played-out. */
-		asDisplayedRundownDuration?: number
-		/** This is the complete duration of the rundown: as planned for the unplayed content, and as-run for the played-out, but ignoring unplayed/unplayable parts in order */
-		asPlayedRundownDuration?: number
-		/** this is the countdown to each of the parts relative to the current on air part. */
-		partCountdown?: {
-			[key: string]: number
-		}
-		/** The calculated durations of each of the Parts: as-planned/as-run depending on state. */
-		partDurations?: {
-			[key: string]: number
-		}
-		/** The offset of each of the Parts from the beginning of the Rundown. */
-		partStartsAt?: {
-			[key: string]: number
-		}
-		/** Same as partStartsAt, but will include display duration overrides
-		 *  (such as minimal display width for an Part, etc.).
-		 */
-		partDisplayStartsAt?: {
-			[key: string]: number
-		}
-		/** Same as partDurations, but will include display duration overrides
-		 * (such as minimal display width for an Part, etc.).
-		 */
-		partDisplayDurations?: {
-			[key: string]: number
-		}
-		/** As-played durations of each part. Will be 0, if not yet played.
-		 * Will be counted from start to now if currently playing.
-		 */
-		partPlayed?: {
-			[key: string]: number
-		}
-		/** Expected durations of each of the parts or the as-played duration,
-		 * if the Part does not have an expected duration.
-		 */
-		partExpectedDurations?: {
-			[key: string]: number
-		}
-		/** Remaining time on current part */
-		remainingTimeOnCurrentPart?: number | undefined
-		/** Current part will autoNext */
-		currentPartWillAutoNext?: boolean
-		/** Current time of this calculation */
-		currentTime?: number
-		/** Was this time context calculated during a high-resolution tick */
-		isLowResolution: boolean
-	}
-
-	/**
-	 * This are the properties that will be injected by the withTiming HOC.
-	 * @export
-	 * @interface InjectedROTimingProps
-	 */
-	export interface InjectedROTimingProps {
-		timingDurations: RundownTimingContext
-	}
-}
 
 const TIMING_DEFAULT_REFRESH_INTERVAL = 1000 / 60 // the interval for high-resolution events (timeupdateHR)
 const LOW_RESOLUTION_TIMING_DECIMATOR = 15 // the low-resolution events will be called every
@@ -596,347 +501,3 @@ export const RundownTimingProvider = withTracker<
 		}
 	}
 )
-
-export type TimingFilterFunction = (durations: RundownTiming.RundownTimingContext) => any
-
-export interface WithTimingOptions {
-	isHighResolution?: boolean
-	filter?: TimingFilterFunction | string | (string | number)[]
-}
-export type WithTiming<T> = T & RundownTiming.InjectedROTimingProps & { children?: React.ReactNode }
-type IWrappedComponent<IProps, IState> = new (props: WithTiming<IProps>, state: IState) => React.Component<
-	WithTiming<IProps>,
-	IState
->
-
-/**
- * Wrap a component in a HOC that will inject a the timing context as a prop. Takes an optional options object that
- * allows a high timing resolution or filtering of the changes in the context, so that the child component only
- * re-renders when a change to the filtered value happens.
- * The options object can also be replaced with an options generator function that will take the incoming props
- * as an argument and produce a {WithTimingOptions} object
- * @export
- * @template IProps The props interface of the child component
- * @template IState The state interface of the child component
- * @param  {(WithTimingOptions | ((props: IProps) => WithTimingOptions))} [options] The options object or the options object generator
- * @return (WrappedComponent: IWrappedComponent<IProps, IState>) =>
- * 		new (props: IProps, context: any ) => React.Component<IProps, IState>
- */
-export function withTiming<IProps, IState>(
-	options?: WithTimingOptions | ((props: IProps) => WithTimingOptions)
-): (
-	WrappedComponent: IWrappedComponent<IProps, IState>
-) => new (props: IProps, context: any) => React.Component<IProps, IState> {
-	let expandedOptions: WithTimingOptions = {
-		isHighResolution: false,
-		...(typeof options === 'function' ? {} : options),
-	}
-
-	return (WrappedComponent) => {
-		return class WithTimingHOCComponent extends React.Component<IProps, IState> {
-			static contextTypes = {
-				durations: PropTypes.object.isRequired,
-			}
-
-			filterGetter: (o: any) => any
-			previousValue: any = undefined
-			isDirty: boolean = false
-
-			constructor(props, context) {
-				super(props, context)
-
-				if (typeof options === 'function') {
-					expandedOptions = {
-						...expandedOptions,
-						...options(this.props),
-					}
-				}
-
-				if (typeof expandedOptions.filter === 'function') {
-					this.filterGetter = expandedOptions.filter
-				} else if (expandedOptions.filter) {
-					this.filterGetter = _.property(expandedOptions.filter as string)
-				}
-			}
-
-			componentDidMount() {
-				window.addEventListener(
-					expandedOptions.isHighResolution ? RundownTiming.Events.timeupdateHR : RundownTiming.Events.timeupdate,
-					this.refreshComponent
-				)
-			}
-
-			componentWillUnmount() {
-				window.removeEventListener(
-					expandedOptions.isHighResolution ? RundownTiming.Events.timeupdateHR : RundownTiming.Events.timeupdate,
-					this.refreshComponent
-				)
-			}
-
-			refreshComponent = () => {
-				if (!this.filterGetter) {
-					this.forceUpdate()
-				} else {
-					const buf = this.filterGetter(this.context.durations || {})
-					if (this.isDirty || !_.isEqual(buf, this.previousValue)) {
-						this.previousValue = buf
-						this.isDirty = false
-						this.forceUpdate()
-					}
-				}
-			}
-
-			render() {
-				const durations: RundownTiming.RundownTimingContext = this.context.durations
-
-				// If the timing HOC is supposed to be low resolution and we are rendering
-				// during a high resolution tick, the WrappedComponent will render using
-				// a RundownTimingContext that has not gone through the filter and thus
-				// previousValue may go out of sync.
-				// To bring it back to sync, we mark the component as dirty, which will
-				// force an update on the next low resoluton tick, regardless of what
-				// the filter says.
-				if (this.filterGetter && durations.isLowResolution !== !expandedOptions.isHighResolution) {
-					this.isDirty = true
-				}
-
-				return <WrappedComponent {...this.props} timingDurations={durations} />
-			}
-		}
-	}
-}
-
-interface IPartCountdownProps {
-	partId?: PartId
-	hideOnZero?: boolean
-}
-
-/**
- * A presentational component that will render a countdown to a given Part
- * @class PartCountdown
- * @extends React.Component<WithTiming<IPartCountdownProps>>
- */
-export const PartCountdown = withTiming<IPartCountdownProps, {}>()(
-	class PartCountdown extends React.Component<WithTiming<IPartCountdownProps>> {
-		render() {
-			return (
-				<span>
-					{this.props.partId &&
-						this.props.timingDurations &&
-						this.props.timingDurations.partCountdown &&
-						this.props.timingDurations.partCountdown[unprotectString(this.props.partId)] !== undefined &&
-						(this.props.hideOnZero !== true ||
-							this.props.timingDurations.partCountdown[unprotectString(this.props.partId)] > 0) &&
-						RundownUtils.formatTimeToShortTime(
-							this.props.timingDurations.partCountdown[unprotectString(this.props.partId)]
-						)}
-				</span>
-			)
-		}
-	}
-)
-
-export const AutoNextStatus = withTiming<{}, {}>({
-	filter: 'currentPartWillAutoNext',
-	isHighResolution: true,
-})(
-	class AutoNextStatus extends React.Component<WithTiming<{}>> {
-		render() {
-			return this.props.timingDurations.currentPartWillAutoNext ? (
-				<div className="rundown-view__part__icon rundown-view__part__icon--auto-next"></div>
-			) : (
-				<div className="rundown-view__part__icon rundown-view__part__icon--next"></div>
-			)
-		}
-	}
-)
-
-const SPEAK_ADVANCE = 500
-
-interface IPartRemainingProps {
-	currentPartInstanceId: PartInstanceId | null
-	hideOnZero?: boolean
-	className?: string
-	heavyClassName?: string
-	speaking?: boolean
-}
-
-// global variable for remembering last uttered displayTime
-let prevDisplayTime: number | undefined = undefined
-
-/**
- * A presentational component that will render a countdown to the end of the current part
- * @class CurrentPartRemaining
- * @extends React.Component<WithTiming<{}>>
- */
-export const CurrentPartRemaining = withTiming<IPartRemainingProps, {}>({
-	isHighResolution: true,
-})(
-	class CurrentPartRemaining extends React.Component<WithTiming<IPartRemainingProps>> {
-		render() {
-			const displayTimecode = this.props.timingDurations.remainingTimeOnCurrentPart
-			return (
-				<span
-					className={ClassNames(
-						this.props.className,
-						Math.floor((displayTimecode || 0) / 1000) > 0 ? this.props.heavyClassName : undefined
-					)}>
-					{RundownUtils.formatDiffToTimecode(displayTimecode || 0, true, false, true, false, true, '', false, true)}
-				</span>
-			)
-		}
-
-		speak(displayTime: number) {
-			let text = '' // Say nothing
-
-			navigator.vibrate([400, 300, 400, 300, 400])
-
-			switch (displayTime) {
-				case -1:
-					text = 'One'
-					break
-				case -2:
-					text = 'Two'
-					break
-				case -3:
-					text = 'Three'
-					break
-				case -4:
-					text = 'Four'
-					break
-				case -5:
-					text = 'Five'
-					break
-				case -6:
-					text = 'Six'
-					break
-				case -7:
-					text = 'Seven'
-					break
-				case -8:
-					text = 'Eight'
-					break
-				case -9:
-					text = 'Nine'
-					break
-				case -10:
-					text = 'Ten'
-					break
-			}
-			// if (displayTime === 0 && prevDisplayTime !== undefined) {
-			// 	text = 'Zero'
-			// }
-
-			if (text) {
-				SpeechSynthesiser.speak(text, 'countdown')
-			}
-		}
-
-		vibrate(displayTime: number) {
-			navigator.vibrate([400, 300, 400, 300, 400])
-
-			switch (displayTime) {
-				case 0:
-					navigator.vibrate([500])
-				case -1:
-				case -2:
-				case -3:
-					navigator.vibrate([250])
-			}
-		}
-
-		act() {
-			// Note that the displayTime is negative when counting down to 0.
-			let displayTime = this.props.timingDurations.remainingTimeOnCurrentPart || 0
-
-			if (displayTime === 0) {
-				// do nothing
-			} else {
-				displayTime += SPEAK_ADVANCE
-				displayTime = Math.floor(displayTime / 1000)
-			}
-
-			if (prevDisplayTime !== displayTime) {
-				if (this.props.speaking) {
-					this.speak(displayTime)
-				}
-
-				this.vibrate(displayTime)
-
-				prevDisplayTime = displayTime
-			}
-		}
-
-		componentDidUpdate(prevProps: WithTiming<IPartRemainingProps>) {
-			if (this.props.currentPartInstanceId !== prevProps.currentPartInstanceId) {
-				prevDisplayTime = undefined
-			}
-			this.act()
-		}
-	}
-)
-
-interface ISegmentDurationProps {
-	partIds: PartId[]
-}
-
-/**
- * A presentational component that will render a counter that will show how much content
- * is left in a segment consisting of given parts
- * @class SegmentDuration
- * @extends React.Component<WithTiming<ISegmentDurationProps>>
- */
-export const SegmentDuration = withTiming<ISegmentDurationProps, {}>()(
-	class SegmentDuration extends React.Component<WithTiming<ISegmentDurationProps>> {
-		render() {
-			if (
-				this.props.partIds &&
-				this.props.timingDurations.partExpectedDurations &&
-				this.props.timingDurations.partPlayed
-			) {
-				let partExpectedDurations = this.props.timingDurations.partExpectedDurations
-				let partPlayed = this.props.timingDurations.partPlayed
-
-				const duration = this.props.partIds.reduce((memo, partId) => {
-					const pId = unprotectString(partId)
-					return partExpectedDurations[pId] !== undefined
-						? memo + Math.max(0, partExpectedDurations[pId] - (partPlayed[pId] || 0))
-						: memo
-				}, 0)
-
-				return (
-					<span className={duration < 0 ? 'negative' : undefined}>
-						{RundownUtils.formatDiffToTimecode(duration, false, false, true, false, true, '+')}
-					</span>
-				)
-			}
-
-			return null
-		}
-	}
-)
-
-/**
- * Computes the actual (as-played fallbacking to expected) duration of a segment, consisting of given parts
- * @export
- * @param  {RundownTiming.RundownTimingContext} timingDurations The timing durations calculated for the Rundown
- * @param  {Array<string>} partIds The IDs of parts that are members of the segment
- * @return number
- */
-export function computeSegmentDuration(
-	timingDurations: RundownTiming.RundownTimingContext,
-	partIds: PartId[],
-	display?: boolean
-): number {
-	let partDurations = timingDurations.partDurations
-
-	if (partDurations === undefined) return 0
-
-	return partIds.reduce((memo, partId) => {
-		const pId = unprotectString(partId)
-		const partDuration =
-			(partDurations ? (partDurations[pId] !== undefined ? partDurations[pId] : 0) : 0) ||
-			(display ? Settings.defaultDisplayDuration : 0)
-		return memo + partDuration
-	}, 0)
-}
