+++ meteor/client/ui/SegmentTimeline/SegmentTimeline.tsx
@@ -12,7 +12,10 @@ import { SegmentUi, PartUi, IOutputLayerUi, PieceUi } from './SegmentTimelineCon
 import { TimelineGrid } from './TimelineGrid'
 import { SegmentTimelinePart } from './SegmentTimelinePart'
 import { SegmentTimelineZoomControls } from './SegmentTimelineZoomControls'
-import { SegmentDuration, PartCountdown, RundownTiming, CurrentPartRemaining } from '../RundownView/RundownTiming'
+import { SegmentDuration } from '../RundownView/RundownTiming/SegmentDuration'
+import { PartCountdown } from '../RundownView/RundownTiming/PartCountdown'
+import { RundownTiming } from '../RundownView/RundownTiming/RundownTiming'
+import { CurrentPartRemaining } from '../RundownView/RundownTiming/CurrentPartRemaining'
 
 import { RundownUtils } from '../../lib/rundown'
 import { Translated } from '../../lib/ReactMeteorData/ReactMeteorData'
@@ -32,11 +35,11 @@ import * as Zoom_Out_MouseOut from './Zoom_Out_MouseOut.json'
 // @ts-ignore Not recognized by Typescript
 import * as Zoom_Out_MouseOver from './Zoom_Out_MouseOver.json'
 import { LottieButton } from '../../lib/LottieButton'
-import { PartNote, NoteType, SegmentNote } from '../../../lib/api/notes'
+import { NoteType, SegmentNote } from '../../../lib/api/notes'
 import { getAllowSpeaking } from '../../lib/localStorage'
 import { showPointerLockCursor, hidePointerLockCursor } from '../../lib/PointerLockCursor'
 import { Settings } from '../../../lib/Settings'
-import { MAGIC_TIME_SCALE_FACTOR, RundownViewEvents, IContextMenuContext } from '../RundownView'
+import { RundownViewEvents, IContextMenuContext } from '../RundownView'
 import { literal, unprotectString } from '../../../lib/lib'
 import { SegmentId } from '../../../lib/collections/Segments'
 import { PartId } from '../../../lib/collections/Parts'
